wcn36xx: do not use test_bit
After running sparse we got couple of warnings on
test_bit function. The problem is that test_bit function
receives position of the bit, but wcn36xx requires
the flag comparison. sparse yields that when calling
test_bit the position of the bit was out of the long boundary.

As a fix compare if flag is set instead of checking bit by
position.

Signed-off-by: Eugene Krasnikov <k.eugene.e@gmail.com>