wcn36xx: fix the mesh STA for rejoin
The firmware has triggered the QoS NULL frame to monitor the links
amongst peer mesh STAs. Once one of the peer mesh STAs is leaving or
out of the range and no ACK received from the peer mesh STA, the
firmware will automatically delete the peer mesh STA.

After deleting the mesh STA and adding the mesh STA back with the
HAL command return no error still won't work. The unicast data frame
is not able to transmit out from the mesh STA. So the only way is
force the mesh STA to use bss STA index instead of the STA index.

Signed-off-by: Chun-Yeow Yeoh <yeohchunyeow@cozybit.com>
Signed-off-by: Jason Mobarak <jam@cozybit.com>