wcn36xx: wake queues less frequently
Once the tx queues fill, the hardware will currently start
disabling queues and restarting them on every packet.  This causes
a lot of context switches, so we can be a little more efficient
here by maintaining a pending count and only waking the queues
when the count is sufficiently low.

Signed-off-by: Bob Copeland <me@bobcopeland.com>