wcn36xx: implement software encryption workaround
When software encryption is enabled, send frames encrypted,
but remove the protected bit.  On the receiver, check if we
have a key for the transmitter, and if so, set the protected
bit so the stack can decrypt it.

This is to workaround a problem where wcn hardware filters
any RX frames for which it has no key in hardware.

Signed-off-by: Bob Copeland <me@bobcopeland.com>